+++ e2e/commands/status.e2e.js
@@ -168,6 +168,31 @@ describe('bit status command', function () {
       expect(output.includes('There are no new components')).to.be.true;
     });
   });
+  describe('when a component is exported, modified, committed and then exported again', () => {
+    let output;
+    before(() => {
+      helper.reInitLocalScope();
+      helper.createComponentBarFoo();
+      helper.addComponentBarFoo();
+      helper.commitComponentBarFoo();
+      helper.reInitRemoteScope();
+      helper.addRemoteScope();
+      helper.exportComponent('bar/foo');
+      helper.createComponentBarFoo("module.exports = function foo() { return 'got foo v2'; };"); // modify the component
+      helper.commitComponentBarFoo();
+      helper.exportComponent('bar/foo');
+      output = helper.runCmd('bit status');
+    });
+    it('should not display that component as modified', () => {
+      expect(output.includes('There are no modified components')).to.be.true;
+    });
+    it('should not display that component as staged', () => {
+      expect(output.includes('There are no staged components')).to.be.true;
+    });
+    it('should not display that component as new', () => {
+      expect(output.includes('There are no new components')).to.be.true;
+    });
+  });
   describe('when a component is imported', () => {
     let output;
     before(() => {